+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/policy/PolicyCheckerService.java
@@ -29,14 +29,16 @@ import java.util.stream.Stream;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.eclipse.edc.policy.model.AndConstraint;
 import org.eclipse.edc.policy.model.AtomicConstraint;
 import org.eclipse.edc.policy.model.Constraint;
-import org.eclipse.edc.policy.model.MultiplicityConstraint;
 import org.eclipse.edc.policy.model.Operator;
+import org.eclipse.edc.policy.model.OrConstraint;
 import org.eclipse.edc.policy.model.Permission;
 import org.eclipse.edc.policy.model.Policy;
+import org.eclipse.edc.policy.model.XoneConstraint;
 import org.eclipse.tractusx.irs.data.StringMapper;
-import org.jetbrains.annotations.NotNull;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.util.UriUtils;
 
@@ -49,6 +51,10 @@ import org.springframework.web.util.UriUtils;
 public class PolicyCheckerService {
 
     private final AcceptedPoliciesProvider policyStore;
+    @Value("${irs-edc-client.catalog.policies.acceptedRightOperands:active}")
+    private final List<String> acceptedRightOperands;
+    @Value("${irs-edc-client.catalog.policies.acceptedLeftOperands:PURPOSE}")
+    private final List<String> acceptedLeftOperands;
 
     public boolean isValid(final Policy policy) {
         final List<PolicyDefinition> policyList = getAllowedPolicies();
@@ -57,23 +63,20 @@ public class PolicyCheckerService {
         if (getValidStoredPolicyIds().contains("*")) {
             return true;
         }
-        return policy.getPermissions()
-                     .stream()
-                     .anyMatch(permission -> policyList.stream()
-                                                       .anyMatch(allowedPolicy -> isValid(permission, allowedPolicy)));
+        return policy.getPermissions().stream().allMatch(permission -> isValid(permission, policyList));
     }
 
-    @NotNull
     private List<PolicyDefinition> getAllowedPolicies() {
         final List<String> policyIds = getValidStoredPolicyIds();
         final List<PolicyDefinition> allowedPolicies = new ArrayList<>();
-        allowedPolicies.addAll(policyIds.stream().map(policy -> createPolicy("idsc:PURPOSE", policy)).toList());
-        allowedPolicies.addAll(policyIds.stream().map(policy -> createPolicy(policy, "active")).toList());
+        acceptedRightOperands.forEach(rightOperand -> allowedPolicies.addAll(
+                policyIds.stream().map(policy -> createPolicy(policy, rightOperand)).toList()));
+        acceptedLeftOperands.forEach(leftOperand -> allowedPolicies.addAll(
+                policyIds.stream().map(policy -> createPolicy(leftOperand, policy)).toList()));
 
         return allowedPolicies;
     }
 
-    @NotNull
     private List<String> getValidStoredPolicyIds() {
         return policyStore.getAcceptedPolicies()
                           .stream()
@@ -83,27 +86,48 @@ public class PolicyCheckerService {
                           .toList();
     }
 
-    private boolean isValid(final Permission permission, final PolicyDefinition policyDefinition) {
-        return permission.getAction().getType().equals(policyDefinition.getPermissionActionType())
-                && permission.getConstraints().stream().anyMatch(constraint -> isValid(constraint, policyDefinition));
+    private boolean isValid(final Permission permission, final List<PolicyDefinition> policyDefinitions) {
+        final boolean permissionTypesMatch = policyDefinitions.stream()
+                                                              .allMatch(
+                                                                      policyDefinition -> policyDefinition.getPermissionActionType()
+                                                                                                          .equals(permission.getAction()
+                                                                                                                            .getType()));
+        final boolean constraintsMatch = permission.getConstraints()
+                                                   .stream()
+                                                   .allMatch(constraint -> isValid(constraint, policyDefinitions));
+        return permissionTypesMatch && constraintsMatch;
     }
 
-    private boolean isValid(final Constraint constraint, final PolicyDefinition policyDefinition) {
+    private boolean isValid(final Constraint constraint, final List<PolicyDefinition> policyDefinitions) {
         if (constraint instanceof AtomicConstraint atomicConstraint) {
-            return AtomicConstraintValidator.builder()
-                                            .atomicConstraint(atomicConstraint)
-                                            .leftExpressionValue(policyDefinition.getLeftExpressionValue())
-                                            .rightExpressionValue(policyDefinition.getRightExpressionValue())
-                                            .expectedOperator(
-                                                    Operator.valueOf(policyDefinition.getConstraintOperator()))
-                                            .build()
-                                            .isValid();
-        } else if (constraint instanceof MultiplicityConstraint multiplicityConstraint) {
-            return multiplicityConstraint.getConstraints()
-                                         .stream()
-                                         .anyMatch(constraint1 -> isValid(constraint1, policyDefinition));
+            return validateAtomicConstraint(atomicConstraint, policyDefinitions);
+        } else if (constraint instanceof AndConstraint andConstraint) {
+            return andConstraint.getConstraints().stream().allMatch(constr -> isValid(constr, policyDefinitions));
+        } else if (constraint instanceof OrConstraint orConstraint) {
+            return orConstraint.getConstraints().stream().anyMatch(constr -> isValid(constr, policyDefinitions));
+        } else if (constraint instanceof XoneConstraint xoneConstraint) {
+            return xoneConstraint.getConstraints().stream().filter(constr -> isValid(constr, policyDefinitions)).count()
+                    == 1;
         }
         return false;
+    }
+
+    private boolean validateAtomicConstraint(final AtomicConstraint atomicConstraint,
+            final PolicyDefinition policyDefinition) {
+        return AtomicConstraintValidator.builder()
+                                        .atomicConstraint(atomicConstraint)
+                                        .leftExpressionValue(policyDefinition.getLeftExpressionValue())
+                                        .rightExpressionValue(policyDefinition.getRightExpressionValue())
+                                        .expectedOperator(Operator.valueOf(policyDefinition.getConstraintOperator()))
+                                        .build()
+                                        .isValid();
+    }
+
+    private boolean validateAtomicConstraint(final AtomicConstraint atomicConstraint,
+            final List<PolicyDefinition> policyDefinitions) {
+        return policyDefinitions.stream()
+                                .anyMatch(policyDefinition -> validateAtomicConstraint(atomicConstraint,
+                                        policyDefinition));
     }
 
     private PolicyDefinition createPolicy(final String leftExpression, final String rightExpression) {